Changed structure of sync stream regarding sinks.
The sync stream now always has three sinks: for the total number of
sync operations, the number of successful sync operations, and a log
sink. The latter is only active is a path to a log file is provided.

This homogeneous structure simplifies the creation of the stream
because it is no longer necessary to have a special treatment for
different cases (e.g. log file supported or not).