Refactored stream for WebDav operation processing.
The single sync operations are now mapped to data objects that contain
a list of the requests to be executed for them. The stream has then
multiple stages to optionally execute a specific request if it is
required for the operation.

Note: A first approach to use a graph with a cycle to handle an
arbitrary number of requests failed because it was not easily possible
to determine when the stream should terminate.