DavConfig now has a list with modified properties.
When retrieving the last-modified time of a file in future all of
these properties will be checked until one is defined. Background is
that if a custom property name is defined, the property might not be
available for a file on the server that has been uploaded using a
different means. Therefore, if this property is missing, also the
default last-modified property should be checked.