Added CryptOpHandler trait.
This trait extracts special functionality required by an encrypt and
decrypt operation respectively. The idea is to parametrize a
CryptStage with such a handler rather than having concrete sub
classes. One advantage of such an approach is that the handler is
stateless and can exists without a concrete CryptStage object. Thus it
can be used by other components (e.g. to determine the correct file
size) as well.