Changed the creation of noop apply stage.
If the apply mode is set to NONE, an apply stage needs to be created
that does not execute any sync operations. The corresponding
functionality has been moved from Sync to SyncStreamFactory. This is
not fully trivial because it has to be ensured that the source file
provider is properly shutdown.