Changed representation of parsed command line parameters.
It is problematic to remove parameters from the parameters map to
indicate that they have been accessed. This prevents that options can
be queried multiple times - which will be needed in future. Therefore,
ParameterManager.Parameters is now a case class storing the original
parameters map and a set with the keys that have been accessed.