Added ActorSystemLifeCycle trait.
This trait implements common functionality related to the management
of an actor system for command line applications. An actor system (and
related objects) is provided in implicit scope. The application logic
is executed, and finally all resources are released.

Another advantage of this trait is that it improves testing of command
line applications.