Added a conditional CliProcessor that dispatches to a group.
This processor runs a selector processor that yields the name of a
group. The processor representing this group is then executed. This is
similar to the already available conditional processor, but there can
be multiple processors to dispatch to. Use cases are commands that
have their specific sets of parameters.