Added an initial module to process structure-related parameters.
The goal is to parse all the parameters for the structures to be
synced into a single configuration object. The new module defines
corresponding data classes for the structures supported.

The StructureType classes have been moved into this module and renamed
to RoleType, which better fits their purpose.

The module is going to be enhanced with functionality to parse the
command line and extract configuration objects for the structures
taking part in the sync process.