SyncStructureConfig now parsers options related to sync structures.
The class is now responsible of generating a StructureConfig object
from the command line for the source and destination structure. These
objects are going to be added to the central sync configuration. (So
command line processing will be done in a single step, rather than
starting another parse operation as is currently done by
SyncComponentsFactory.

The latter is going to be reworked; there is currently a lot of
redundancy between SyncComponentsFactory and SyncStructureConfig.