Introduced a special exception for failed parameter extractions.
Rather than throwing an IllegalArgumentException, now a new
ParameterExtractionException is contained in the Failure. The
exception can hold a list of failures. Thus, the failure messages are
no longer converted to a single string, but can now be accessed one by
one.

A number of tests had to be adapted to the new exception type.