Added new ParameterParser module.
This module contains functionality related to the parsing of the array
with command line arguments into a map of parameters. It has been
extracted from ParameterManager and slightly modified.

In a first step, the parsing functionality is copied more or less
directly; it is going to be made more generic later on.