Added initial version of ParameterManager.
This module provides a facade for the other services related to CLI
processing. It offers a central function to parse the command line
arguments and to extract a configuration object out of the results.

The current version is functional, but there is room for improvement
in the way failures to parse the command line are handled.