Moved common CLI processing logic to CliActorSystemLifeCycle.
The trait was renamed to better reflect its support for CLI
processing. The whole parsing and extraction of a configuration object
is now done by the trait; the runApp() method is then invoked with the
future result of the configuration.

Unfortunately, this caused a number of adaptions in test classes.