Added some aliases for the options of the OAuth application.
The parameters related to the storage configuration can now be
abbreviated. However, this is only true if they are used in the
context of the OAuth application and not from the main sync
application. Here the options refer to the source or the destination
structure and thus cannot be expressed as alias.