Fixed a regression with reading parameter files.
When moving to the SCLI library to parse the command line options the
ability to load parameter files got lost. (This is an optional step in
SCLI, which needs to be invoked explicitly; unfortunately, this has
been forgotten.) This has been fixed; files referenced on the command
line are now read again.