Use defensive completion of the promise in the OAuth login command.
There were exceptions observed that the promise has already been
completed, which may happen if the request handler is invoked multiple
times. (The exception stack trace was printed on the console, but the
command was executed correctly.) To prevent this, the tryXXX() methods
are now used when completing the promise.