Reworked OAuthStorageService and its implementation.
The functions to load or save specific parts of the OAuth
configuration are in most cases not needed any more; therefore, they
have been removed from the trait and made private in the
implementation.

Usage of the custom CryptService has been replaced by the one provided
by the CloudFiles library.