Added SyncProtocol trait.
The trait defines the protocol used by a sync process to manipulate
data structures. It contains the operations required to iterate over
the elements contained in a structure, to update the destination
structure accordingly (by creating files and folders, updating or
deleting them), and to download files from the source structure.