Simplified the types for sync protocol factories.
The types are no longer generic in the concrete type of the
protocol-specific configuration. The existing design was overly
complex and only caused problems when instantiating concrete
factories. Therefore, now the configuration object is not passed to
the factory function, but it is passed to the factory at creation
time.