Added a new dstID property to SyncOperation.
This property is going to reference the element in the destination
structure affected by a sync operation. It is now evaluated by
ProtocolOperationHandler rather than the srcURI property. It is going
to replace the srcURI and dstURI properties, which are no longer
needed by the classes based on CloudFiles file systems.

By introducing this new property, the switch to file systems can be
prepared without breaking existing sync code.