Switched to implementations based on FileSystems.
Sync now uses SyncProtocols for the source and destination structures
to perform a sync process.

This changes caused a bunch of updates in tests, although the
expectations remain the same. Note that the integration tests now
partly test functionality provided by the CloudFiles library; however,
as they already exist, it is moot to convert them to a pure mocking
approach.

A few tests have been disabled; the functionality they test is going
to be removed in future.