ProtocolOperationHandlerStage now uses a higher parallelism.
Analogously to the original implementation, the stage now uses the
size of the HTTP connection pool as parallelism factor when executing
sync operations asynchronously.

As this may lead to race conditions, when operations are executed in
parallel that depend on each other, logic is implement to keep track
on the operations in progress and to delay specific operations that
can only run after others are completed.