Moved classes related to OAuth into a dedicated top-level package.
The classes used to be located below the http package. As there is no
functionality for sending HTTP requests any more (this is now provided
exclusively by CloudFiles), this package is obsolete. It therefore
makes sense to move the http.oauth package one level up below the sync
package.

SyncAuthConfig.scala has been moved into this package as well.
Although the file contains configuration classes for other
authentication mechanism as well, there is a strong relation to the
other classes in the oauth package.