Added basic logging facilities.
Some changes in the past broke logging, so that no output was
generated. Dependencies to Apache log4j2 have been added and a very
basic logging configuration to re-enable logging. (This configuration
is going to be enhanced later.)

Akka makes use of slf4j for logging. The corresponding bridge from
log4j2 has been added to route the output to the log4j2 system.