Added logging of file attributes causing override actions.
It is sometimes not obvious why a file is overridden during a sync
process. In debug level, SyncStage now outputs the attributes of both
files when the destination is overridden from the source.

To actually enable DEBUG logging from Akka loggers, it is necessary to
set the log level in application.conf (which is INFO per default);
otherwise, log events with a lower level are not passed to the logging
backend.