Enhanced file system conversion in FileSystemProtocolConverter.
The function to convert a sync file to a file system file now expects
an additional parameter that determines whether the file ID should be
converted or not.

During a sync process, the ID is always present, but it may stem from
a different file system (which may cause the conversion to fail).
FileSystemSyncProtocol now distinguishes between the creation of a new
file and an overwrite operation. In the latter case, the ID is present
and needs to be converted; in the former case, it should be ignored.