Made the test configuration compile.
A version clash with the scala-xml module had to be resolved to use
the Scala 3 version of ScalaTest. Then it was necessary to switch to a
new version of the ScalaTest Mockito integration. This caused some
changes in the Mockito API.

Some tests are not successful due to an incompatibility between
ScalaTestWithActorTestKit and Scalatic.