Introduced a new SyncAction case Noop.
This case indicates that no actions are required for a specific
element (since it is the same for both sync structures). The reason
for this new action is that all elements should be passed through the
sync stream, including the ones that need not be manipulated. This
will be necessary for bi-directional sync processes to write a
snapshot of the local structure.