Added (initial) SyncStream module.
This object is going to provide functionality related to the creation
of sync streams and the single components that comprise such streams.
The idea is to have a parameters object that describes the stream;
based on these parameters, the sync stream is then constructed.

The object is going to be extended with functions that create special
stream components, e.g. pre-defined sinks or sources.