Sync now uses functionality of SyncStream to construct the stream.
Sync.createSyncStream() now delegates to the SyncStream module. Some
changes were required to adapt the functions to the different types
used by the module. SyncResult now stores the number of successful and
failed operations and calculates the total number of operations.

Made Sync.createSyncStream() private; it is not called externally.