Implemented a more generic approach to combining sync sinks.
Adding logging to an existing sink is just a special case. SyncStream
now offers a function to combine two arbitrary sinks, ignoring the
materialized value of the second sink. sinkWithLogging() is
implemented using this new function.

The new function is used by Sync to add another Sink to the sync
stream that logs errors to the console. The ugly work-around in
SyncProtocolHolder for logging errors could now be removed.