Added BaseMergeStage trait.
This trait provides common logic for stage implementations that
somehow need to merge two input sources. The logic has been extracted
from MirrorStage and made rather generic.

For the implementation of bidirectional sync processes, further stages
are needed that need to do similar merge operations. Therefore, it
makes sense to separate this logic, so that it can be reused.