Simplified completion checks in MirrorStage.
By using new functionality from the base trait, it is no longer
necessary to store the number of finished sources in MirrorState. This
field was only used in a special case, that one source was completed
before the other. By using the inherited numberOfFinishedSources
function and slightly rewriting the
handleNullElementOnFinishedSource() function, the field is no longer
needed.