Added LocalStateStage class.
This is a stage implementation that compares the elements from a local
folder structure against a recorded state from the last sync process.
That way, additional information about local changes can be provided.

The current implementation contains some redundancy with the
MirrorStage class. This should be dealt with in a later commit.