BaseMergeStage: Changed representation of undefined elements.
Instead of using null references, the merge stages now use the Option
type for elements that can be undefined, e.g. to indicate that a
source is complete or that there is no current element in the merge
state.

This enables more idiomatic expressions and will also be beneficial
for the upcoming bidirectional sync stage.