BaseMergeStage: Made EmitData available externally.
So far, the type could only be accessed within this trait and derived
classes. It is, however, useful - for instance in helper classes - to
access this type from other places as well. Therefore, the class has
been moved as MergeEmitData with a type parameter to the companion
object. EmitData is now an internal type alias.