SyncTypes: Changed the way elements are compared.
The new algorithm sorts elements with a greater level after elements
with a lower level and takes the BFS order of sources into account.
The previous implementation obviously had the same intention, but was
more complex and probably incorrect if there was a larger delta
between the levels of the elements to compare.