Added RemovedFolderConflictHandler class.
This class is responsible for the dealing with remove folder
operations. Such operations are complicated, since they can produce
conflicts. Where a conflict occurs or not, and which operations are
affected, becomes apparent only at the very end of the operation;
therefore, for each operation state information needs to be kept.

RemovedFolderConflictHandler manages this state and updates it when
new elements are processed. It is going to be integrated within
SyncStage, so that removals of folders on both sides can be handled
correctly.