SyncStage: Fixed a problem in syncRemaining().
The current implementation just created elements on the opposite side,
but did not take local change types into account. In fact, the same
cases have to be distinguished as is done by the sync functions for
different elements. This has been fixed by simply delegating to the
corresponding sync functions.