RemovedFolderConflictHandler: Introduced a ResultFunc.
So far, the handler has been initialized with the inlets to pull
after generating a result. When dealing with removals of local
folders, this will no longer be sufficient; then different inlets need
to be pulled, depending on the context in which the handler is called.

The new ResultFunc is a means to parametrize this context. It gives
the caller full control over all the properties of the final result.

The result function used for now is still specific to the use case of
removing remote folders. In needs to be extended for local folders.