Added a Throttle module.
The current approach to throttling does no longer work because the
Noop actions emitted for all unchanged elements. They should not be
subject to throttling, as this slows down a sync stream extremely.

To deal with this issue, a new module has been introduced that
provides a more sophisticated implementation of throttling that only
takes "real" actions into account.