SyncConfig: Extended and restructured the configuration.
The configuration now has an additional property for the time unit for
throttling. The opsPerSecond property has been renamed to opsPerUnit.
The command-line option has been renamed to "throttle".

With this new property, a limit for the supported number of parameters
for the createRepresentation() function was reached. To work around
this, a new nested configuration class, LogConfig, was introduced that
collects the different properties related to logging.