Build: Disabled parallel test execution for integration tests.
There were suddenly test failures in the tests that checked log
output, obviously caused by tests running in parallel and messing up
the expected log output. It is unclear, why this happens now and not
before. The issue is fixed by disabling parallel test execution for
the integration tests - which is not a too bad option.