SyncOperation: Added the deferred property.
This property indicates that the operation is executed in a different
order than the normal element processing order. This is going to be
needed for bidirectional sync processes to correctly handle complex
operations such as folder removals. Then the operations to delete the
elements of the folder cannot be processed before the whole folder has
been processed and no conflicts have been found.