ElementSerializer: Started to rework serialization.
As there are multiple types that require serialization and
deserialization, a new concept was introduced based on a new
SerializationSupport type class. ElementSerializer now offers generic
functions to serialize and deserialize objects for which an instance
of this type class exists.

Currently, an instance has been implemented for FsElement. Further
instances are to follow.