LocalState: Changed the API to obtain sources and sinks.
Introduced a new LocalStateFolder class that contains the information
required to resolve files with local state and offers corresponding
functions. The functions of LocalState that return Source or Sink
objects now expect an object of this type as argument.

This refactoring makes the API a bit leaner (only a single parameter
has to be passed). In addition, useful functionality to access state
files is directly available where it is needed.