Added IgnoreTimeDelta class.
The class represents the delta in the timestamps between to files that
is ignored before these elements are considered different. It also
contains the comparison logic.

The class is going to replace the delta value and comparison
functionality in MirrorSink. Since similar logic needs to be added to
SyncStage, it makes sense to refactor the logic into a separate class.
Also, having a dedicated value type for the ignore-time-delta property
makes the code more readable.

Signed-off-by: oheger <oliver.heger@oliver-heger.de>