SyncStream: Introduced a base class for stream parameters.
The parameters for sync streams will be different to the ones for
mirror streams, but there are a number of properties common to both.
The new class BaseStreamParams manages these. MirrorStreamParams was
adapted accordingly; it now embeds an object of BaseStreamParams, but
exports it, so that the base properties can be queried directly.

Signed-off-by: oheger <oliver.heger@oliver-heger.de>