SyncParameterManager: Extracted a helper class for stream parameters.
The StreamConfig class and the logic to populate it has been moved to
a dedicated module named SyncCliStreamConfig. It is going to be
extended when different kinds of streams (sync, mirror) are supported.
Since SyncParameterManager is already rather complex, it makes sense
to have this logic in a separate module.

Signed-off-by: oheger <oliver.heger@oliver-heger.de>