Build: Updated WireMock dependency from 2.33.2 to 2.34.0.
This was tricky, since the new version of WireMock caused some
integration tests related to logging output to fail. This was caused
by the new WireMock version using SLF4J 2 instead of 1.7.x, which
switched to a different mechanism for its logging adapters. To fix
this, tests now use a different log4j-slf4j bridge than the normal
code.

Signed-off-by: oheger <oliver.heger@oliver-heger.de>