build: Switched from Akka to Apache Pekko.
Upgraded the dependency to CloudFiles from 0.5 to 0.6, which requires
this switch. Continuing using Akka is no longer an option because it
has changed its license.

Since Apache Pekko is a fork of the latest Akka version under the
Apache License, the switch could be done by replacing the
corresponding import statements.

Signed-off-by: Oliver Heger <oliver.heger@oliver-heger.de>