MirrorStage: Fixed a bug in comparing sync elements.
The mirror stage produced invalid sync operations if the last element
on a processing level was removed. In this case, no action to remove
this element was created, but actions to create the elements that were
following in the source structure. This has been fixed by adapting the
comparison of elements to include the current level.

Signed-off-by: Oliver Heger <oliver.heger@oliver-heger.de>